+++ client/go.mod
@@ -1,0 +1,5 @@
+module client
+
+go 1.15
+
+require google.golang.org/grpc v1.31.1
