+++ proto/6.0.1/go.mod
@@ -1,0 +1,9 @@
+module proto/6.0.1
+
+go 1.15
+
+require (
+	github.com/golang/protobuf v1.4.2
+	google.golang.org/grpc v1.31.1
+	google.golang.org/protobuf v1.25.0
+)
