+++ server/packets/6.0/version.go
